# apps/api/src/main/java/com/spellbound/catalogue/CategoryDto.java
package com.spellbound.catalogue;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CategoryDto {
    private UUID id;
    private String slug;
    private String name;
    private String description;
    private UUID parentId;
}