# apps/api/src/main/java/com/spellbound/catalogue/ProductUpdateRequest.java
package com.spellbound.catalogue;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class ProductUpdateRequest {
    private String name;
    private String description;
    private UUID categoryId;
    private ProductStatus status;
    private String attributes;
}