# apps/api/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: Spellbound E-Commerce API
  description: API contract for Spellbound Storefront, Admin Panel, and Backend integration.
  version: v0.1.0
servers:
  - url: http://localhost:8080
    description: Local Development Server
paths:
...
  /api/v1/products:
    get:
      tags:
        - Product
      summary: Get paginated products with optional filters
      operationId: getProducts
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: categoryId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated product list
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductDto'
                  totalElements:
                    type: integer
                    format: int64
                  totalPages:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
    post:
      tags:
        - Product
      summary: Create a new product
      operationId: createProduct
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreateRequest'
      responses:
        '201':
          description: Product created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
  /api/v1/products/{id}:
    get:
      tags:
        - Product
      summary: Get product by ID
      operationId: getProductById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
    put:
      tags:
        - Product
      summary: Update product details
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdateRequest'
      responses:
        '200':
          description: Product updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
    delete:
      tags:
        - Product
      summary: Archive product by ID
      operationId: archiveProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Product archived
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
  /api/v1/products/slug/{slug}:
    get:
      tags:
        - Product
      summary: Get product by slug
      operationId: getProductBySlug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
...
components:
  schemas:
    ProductStatus:
      type: string
      enum:
        - DRAFT
        - ACTIVE
        - ARCHIVED
    SyncStatus:
      type: string
      enum:
        - SYNCED
        - PENDING
        - FAILED
        - WEB_ONLY
    CategoryDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        parentId:
          type: string
          format: uuid
...
    ProductVariantDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        productId:
          type: string
          format: uuid
        sku:
          type: string
        optionValues:
          type: string
        priceCents:
          type: integer
          format: int64
        compareAtCents:
          type: integer
          format: int64
        posExternalId:
          type: string
    ProductVariantCreateRequest:
      type: object
      required:
        - sku
        - optionValues
        - priceCents
      properties:
        sku:
          type: string
        optionValues:
          type: string
        priceCents:
          type: integer
          format: int64
        compareAtCents:
          type: integer
          format: int64
        posExternalId:
          type: string
    ProductDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          $ref: '#/components/schemas/CategoryDto'
        status:
          $ref: '#/components/schemas/ProductStatus'
        attributes:
          type: string
        posExternalId:
          type: string
        lastSyncedAt:
          type: string
          format: date-time
        syncStatus:
          $ref: '#/components/schemas/SyncStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariantDto'
    ProductCreateRequest:
      type: object
      required:
        - slug
        - name
      properties:
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ProductStatus'
        attributes:
          type: string
        posExternalId:
          type: string
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariantCreateRequest'
    ProductUpdateRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/ProductStatus'
        attributes:
          type: string
...